+++ shuup/customer_group_pricing/signal_handers.py
@@ -8,7 +8,9 @@
 from __future__ import unicode_literals
 
 from django.db.models.signals import m2m_changed, post_save, pre_delete
+from django.dispatch import receiver
 
+from shuup.admin.signals import product_copied
 from shuup.core.models import Contact, ContactGroup
 from shuup.core.utils.price_cache import bump_all_price_caches
 from shuup.customer_group_pricing.models import CgpDiscount, CgpPrice
@@ -30,6 +32,15 @@ def handle_contact_group_m2m_changed(sender, instance, **kwargs):
             bump_all_price_caches([instance.shop_id])
         else:
             bump_all_price_caches()
+
+
+@receiver(product_copied, dispatch_uid="customer_group_pricing_product_copied")
+def handle_product_copy(sender, shop, copied, copy, **kwargs):
+    for price in CgpPrice.objects.filter(product=copied, shop=shop):
+        CgpPrice.objects.create(
+            product=copy, shop=shop, group=price.group,
+            price_value=price.price_value
+        )
 
 
 # Bump prices cache when CgpDiscount is changed or deleted