+++ shuup/simple_supplier/__init__.py
@@ -30,5 +30,8 @@ class ShuupSimpleSupplierAppConfig(AppConfig):
         ]
     }
 
+    def ready(self):
+        import shuup.simple_supplier.product_copy_signal_handler  # noqa: F401
+
 
 default_app_config = "shuup.simple_supplier.ShuupSimpleSupplierAppConfig"